backend/requirements.txt
uvicorn
aiogram
fastapi
python-dotenv